add register_flow_pid_for_sock helper function
