+++ pkg/security/ebpf/c/include/hooks/procfs.h
@@ -6,6 +6,7 @@
 #include "constants/offsets/netns.h"
 #include "constants/offsets/network.h"
 #include "helpers/filesystem.h"
+#include "helpers/network/flow.h"
 #include "helpers/utils.h"
 
 static __attribute__((always_inline)) void cache_file(struct dentry *dentry, u32 mount_id) {
@@ -120,10 +121,6 @@ int hook_path_get(ctx_t *ctx) {
 
     struct path *p = (struct path *)CTX_PARM1(ctx);
     struct file *sock_file = (void *)p - f_path_offset;
-    struct pid_route_t route = {};
-    struct pid_route_entry_t value = {};
-    value.pid = *procfs_pid;
-    value.type = PROCFS_ENTRY;
 
     struct socket *socket;
     bpf_probe_read(&socket, sizeof(socket), &sock_file->private_data);
@@ -131,44 +128,8 @@ int hook_path_get(ctx_t *ctx) {
         return 0;
     }
 
-    struct sock *sk = get_sock_from_socket(socket);
-    if (sk == NULL) {
-        return 0;
-    }
+    register_flow_pid_for_sock(get_sock_from_socket(socket), *procfs_pid);
 
-    route.netns = get_netns_from_sock(sk);
-    if (route.netns == 0) {
-        return 0;
-    }
-
-    route.port = get_skc_num_from_sock_common((void *)sk);
-    if (route.port == 0) {
-        // without a port we can't do much, leave early
-        return 0;
-    }
-    route.l4_protocol = get_protocol_from_sock(sk);
-    u16 family = get_family_from_sock_common((void *)sk);
-    if (family == AF_INET6) {
-        bpf_probe_read(&route.addr, sizeof(u64) * 2, &sk->__sk_common.skc_v6_rcv_saddr);
-        bpf_map_update_elem(&flow_pid, &route, &value, BPF_ANY);
-
-        // This AF_INET6 socket might also handle AF_INET traffic, store a mapping to AF_INET too
-        family = AF_INET;
-    }
-    if (family == AF_INET) {
-        bpf_probe_read(&route.addr, sizeof(sk->__sk_common.skc_rcv_saddr), &sk->__sk_common.skc_rcv_saddr);
-        bpf_map_update_elem(&flow_pid, &route, &value, BPF_ANY);
-    } else {
-        // ignore unsupported traffic for now
-        return 0;
-    }
-
-#if defined(DEBUG_NETNS)
-    bpf_printk("path_get netns: %u", route.netns);
-    bpf_printk("         skc_num:%d", htons(route.port));
-    bpf_printk("         skc_rcv_saddr:%x", route.addr[0]);
-    bpf_printk("         pid:%d", pid);
-#endif
     return 0;
 }
